refactor(accesscore): split cell.go into cell_{init,routes}.go
cell.go      (~173 lines) — package doc, interface checks, Option funcs, struct, NewAccessCore
cell_init.go (~189 lines) — Init, initValidate, initSlices, initRbacAssign, LifecycleHooks
cell_routes.go (~112 lines) — HealthCheckers, TokenVerifier, Authorizer, RegisterRoutes, RegisterSubscriptions
